05_Post_Intervention_Measurement/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
feedparser>=6.0.10
openai>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
python-dateutil>=2.8.2